Modified NGram.py for Negative Intervals on Voice Crossing
Previously, the lowest Note of Interval objects given to NGram() were
always interepreted as the lowest voice. This is not the professors'
desired behaviour--they want to know when a voice crossing occurs, so
we must show this as a negative interval.

Adjustments to the rest of the program are still required, especially
in unit tests.