Finish the "Show" UI Component
Interval and n-gram results are no longer printed automatically, but rather
when a user uses the "show" command. Output formatting options are
decided at on the fly, as is the filename. This draws to a close the current
round of output formatting improvements, and closes #19.

Also in this commit is the ground-work for analyzing multiple files. It's
already possible, and statistics are cumulative, but vis will not yet
take a directory name as input.