Modified NGram.__repr__()
I rewrote the __repr__() method so it follows the python guidelines
that suggest its return value should be sufficient to create a new
object with the same values. Of course, music21 doesn't do this, so
it's a little bit jarring, but we'll see how it goes. This works
toward Issue #34