Fixed Automated Tests after Vertical_Interval_Statistics Internal Changes
There was a change in the V_I_S internal ngram dict; it maintained "API
compatibility" (ha!) but broke the automated tests. I've now fixed most of
those tests, which included adding a new method to V_I_S that will again
output the contents of the dict in an easily human-readable format.