Added GUI Ability to Choose Parts to Analyze
Fixes #46. There's a new voice-selection QDialog window that optionally
displays the part names, and allows users to choose how they want to
specify which parts to analyze.

The new QDialog window is integrated with the GUI version of
analyze_this().

There is a new property in VIS_Settings() called "howToChooseVoices".
It corresponds to the new GUI option, "How to Choose Voices to Analyze."