Write NGram.make_from_str() and Clean Up Automated Tests
make_from_str() takes the output of str(NGram) and creates a new NGram
instance with the intervals specified. This is likely the easiest way for
users to tell us what they want to find--for instance, if they were to
invent an n-gram and ask us to find its inversions (as issue #55).

I also significantly cleaned up (segmented) the automated test for NGram.