Add Test Confirmation for make_basso_seguente()
I guess this is a cop-out, but for such a short method, it doesn't seem
worth the effort of making extensive automated tests when it's much
faster to see the method works by visual confirmation. So I offer this
visual confirmation, which is the result of running the doctest example
included with make_basso_seguente().