Improve output_LilyPond Automated Tests
I divided many of the unit tests into independent test cases to keep
track of which ones are failing more effectively. Of course, they all
pass for now, except one of the integration tests that used to work.

I also improved duration_to_lily() by removing code that my tests indicate
may be unused. Obviously, I'm not 100% sure it's never used, but if it is,
it's in highly complex cases and I need to figure it out and write tests
for them.