Rewrote NGram.make_from_str
I was getting some weird errors when I tried to use this method for
JSON stuff, so I rewrote it in (what I thought was) a simpler form.
Also, I thought that the 14th and 15th tests may not actually reflect
use cases of the method... why do we want to consider such strange
strings as valid N-Grams?