Working "repetition-computation" script
Usage is `python compute_repetition.py /path/to/music_file.extension`.
The results are written in the generated file results.csv. I had to make
some minor changes to Analyzer for this to fully work; Nothing crazy though,
just the one line in Analyzer.run_analysis() which will need to change in the
long run.