GUI Parts List Uses Scroll Bars
It used to be that pieces with lots of parts would make the whole vis
window expand---even if it meant making part of the window go off the
screen. Now, if there are too many parts, a scroll bar appears to help
you see all the parts, and to avoid making the window expand.

Closes #213