Revise "About" and "WorkflowManager" API Docs
Especially in adding better documentation of how to use the
WorkflowManager with len() and indexing and so on.

This also means that giving IndexedPiece objects to WorkflowManager is an
undocumented feature; I only left it because it's important for testing.
(cherry picked from commit 3248b796923f63f4cce710c4049fde14713cef49)