CHANGE: default of 'continuer' to 'dynamic quality'
The previous default was for the 'continuer' to be set to label
unisons that arise from stationary lower voices with '_'. Now the
default is 'dynamic quality' which sets the label to 'P1' or '1'
depending on whether 'interval quality' is set to True or False
respectively. The user can still override it with a custom string.
Some tests were modified to accomodate this change.

Related to #309.