Fixed: continuer now returns to 'dynamic quality' when it should.
When the continuer was left to it's default of 'dynamic quality',
it would change to 'P1' or '1' after running an experiment. Now
that change still happens for experiments, but after each experiment
the continuer now gets reset to 'dynamic quality' if that's what
it's setting was before the experiment.

Related to #309.