Fixed: order of operations in workflow.py which corrects offset abstractions
Previously the notes and intervals indexers were called together before
run_off_rep() and the ngram indexer. Now the order is notes, run_off_rep,
intervals, ngram indexer, and the rest is unchanged. The purpose of this is
to correct the calculation of horizontal intervals in n-grams when filtering
by offset. The fix works but the tests still need to be rewritten to
accomodate these changes.

Related to #310