Tests: integration test on workflow.py now correspond to new workflow
Putting the interval indexer after the offset indexer in the workflow
causes the continuer to no get used when pieces are filtered by offset.
This corrects the integration tests of the continuer settings by running
analyses without filtering by offset. Three csv files of expected
results are added. Now the Kyrie_short.krn piece in the test corpus is
unused but it is being left in because it may be used later.

Related to #310