New: support for fake suspensions in dissonance.py
"Fake" suspensions are now detected and are labelled with the _SUSP_FAKE_LABEL
which is "FSUS" by default. The first of the two fourths receives this label
which means that the second fourth should still receive a regular suspension
label of "SUSP". We may later wish to consolidate these two labels into one.