Fixed: detection of dissonant fifths as neighbor tones in dissonance.py
Now neighbor tones that are dissonant fifths (either diminished or augmented)
are detected and marked accordingly. In some cases one may wish to call a
perfect fifth a neighbor note but these are not detected as the fifth in
question is consonant. But this is probably only an issue in two voices,
because with more voices the neighbor tone will inevitably be dissonant against
another voice.