Fixed: detection of dissonant fifths as passing tones in dissonance.py
Dissonant fifths can now be detected as passing tones. This will probably
make the most impact in two-voice textures where there are not additional
parts that the passing tones would get detected against. Also, this should
allow for the labelling of two consecutive passing tones in a single voice
such as the line G-F-E-D over a sustained B. The F and E will now each
be labelled as passing tones.