New: Add 'horiz_attach_later' Setting for HorizontalIntervalIndexer
In VIS 1, all horizontal intervals would be recorded at the offset of the later
note in the interval. In early VIS 2, this was changed to the offset of the
earlier note in the interval. Now there is a choice, defaulting to the earlier,
which more closely reflects the behaviour of nearly every other analyzer.

Related to #298 and #327