Fixed: LilyPondExperimenter.required_score_type = 'stream.Score'
As this class changed from an Indexer to an Experimenter, I lost the
functionality of automatically importing the Score for outputlilypond. This
doesn't require a change in IndexedPiece---just in the Experimenter.

I also added a note in the Experimenter and TemplateExperimenter, saying it's
possible to set required_score_type.

Related to #330