New: Importing no longer uses/stores cached pickle files.
Previously, every import would store its Score as a pickle, or try to load the
pickled Score from a previous load. This isn't likely what we want for VIS,
since we will implement our own cross-user caching, and the time it takes to
store as a pickle is wasted anyway.