Fixed: WorkflowManager.output() acts differently when "count frequency" is False
This is for table-like outputs only: CSV, Stata, Excel, and HTML.

Previously, output() would fail if called with table-like output formats but the
"count frequency" setting were False. Now, it outputs the unaggregated results,
one file per piece, quasi-LilyPond-style.

This is required for "moment-by-moment" output in the Counterpoint Web App.