Added .stripTies() to stream importation
Using .stripTies() rewrites ties as notes with the appropriate duration and beatstrength. Instead of being two events, two tied notes become just one event as most music analysts would probably consider them to be. This slows down the importation somewhat and eliminates measures but they may be able to be retained with the retainContainers=False argument.