New Measure Indexer.
Added measure indexer to metre.py so that we can readily figure out what measure any given offset is in. It only actually indexes the measures of the part stream passed so it assumes that all parts have an equal number of measures and change measures at the same time.