New: measure indexer in metre.py
Replaces old indexer which was just commented out anyway. Now it
indexes the same way the other indexers do. An alternative would
have been to use music21's measureTemplate() method but that
would make this an outlier among the indexers and wouldn't allow
for the future workflow that combines all stream indexers which
I plan to eventually implement.