Refactor: rewrite of ngram indexer.
Currently works but returns all combinations analyzed as different
columns in the same dataframe, rather than just one combination
per dataframe as before. So this change is incompatible with the
tests and with the previous way of calling the ngram indexer and
so it cannot currently be called by the Workflow Manager.