New: deprecate ngram.py, replace with new_ngram.py.
Deprecates current ngram indexer and replaces it with new_ngram
indexer. new_ngram has all of the functionality of the previous
ngram indexer and adds some convenient settings, analysis
possibilities, and output control. Notably, all the voice
combinations examined in a given piece can be returned as a
dataframe with as many columns as there are voice combinations
the user is identifying ngrams in.

closes 364, 360, 359, 352, 318
ref 343, 274