Refactor: handling of self._score in Indexer superclass
Removes problematic and inefficient chained indexing when converting
from a dataframe to a list of series according the indexer's
required_score_type. Also retains level one of the column names
when converting from dataframe to list of series instead of defaulting
to a list of ascending integers as strings. Also updates test to
reflect this.
Related to #440